% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Current time}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{Language either "es" (Spanish) or "en" (English)}
}
\value{
A character string
}
\description{
Returns a sentence with the current time
}
\examples{
what_time()
}
