% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempo.R
\name{what_time}
\alias{what_time}
\title{Da la hora}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{idioma , o "es" o "en"}
}
\value{
Un frase
}
\description{
Da la hora
}
\examples{
what_time()
what_time("es")
}
