% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiemp.R
\name{what_time}
\alias{what_time}
\title{Da la hora}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{Idioma, o "es" o "en"}
}
\value{
Una frase
}
\description{
Da la hora
}
\examples{
what_time()
what_time("en")
}
